pyproject.toml
[project]
name = "agentic-empathy-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "bitsandbytes>=0.49.1",
    "chromadb>=1.4.1",
    "elevenlabs>=2.33.0",
    "fastapi>=0.135.1",
    "langchain-chroma>=1.1.0",
    "langchain-openai>=1.1.7",
    "matplotlib>=3.10.8",
    "neo4j>=6.1.0",
    "numpy>=2.4.1",
    "openai>=2.15.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "peft>=0.18.1",
    "pygame>=2.6.1",
    "python-dotenv>=1.2.1",
    "sounddevice>=0.5.5",
    "soundfile>=0.13.1",
    "torch>=2.10.0",
    "transformers>=4.57.6",
    "uvicorn[standard]>=0.40.0",
    "websockets>=16.0",
]
